add files regarding channel mode in a directory
