+++ Makefile
@@ -21,7 +21,11 @@ SRCS		= main.cpp ManageServer.cpp ManageServerUtils.cpp parsing.cpp utils.cpp \
 				commands/oper.cpp	\
 				commands/user.cpp 	\
 				commands/kill.cpp	\
-				commands/motd.cpp
+				commands/motd.cpp	\
+				commands/channel_modes/key.cpp 		\
+				commands/channel_modes/operator.cpp \
+				commands/channel_modes/secret.cpp 	\
+				commands/channel_modes/topic.cpp 	\
 
 DIR_SRCS	= srcs/
