+++ srcs/commands/mode.cpp
@@ -107,14 +107,7 @@ In a channel that does not have this mode enabled, anyone may set the topic of t
 
 */
 
-struct mode_struct
-{
-	std::string	target;
-	std::string	mode;
-	std::string	params;
-};
-
-static void	broadcastToAllChannelMembers(Server *server, Channel &channel, std::string reply)
+void	broadcastToAllChannelMembers(Server *server, Channel &channel, std::string reply)
 {
 	std::map<std::string, Client>::iterator member = channel.getClientList().begin();
 	
@@ -157,154 +150,6 @@ static void	fillModeInfos(mode_struct &mode_infos, std::string command)
 
 	// PARAM
 	mode_infos.params = command.substr(0);
-}
-
-static void	operatorChannelMode(Server *server, mode_struct mode_infos, int const client_fd, std::string str)
-{
-	std::map<const int, Client>::iterator it_client = server->getClients().find(client_fd);
-	std::map<std::string, Channel>::iterator it_channel_target = server->getChannels().find(mode_infos.target);
-	
-	if (mode_infos.params.empty() == true)
-		return ;
-
-	// checker que le param est un user IRC,puis user du channel
-	std::map<const int, Client>::iterator it_user_target = server->getClients().begin();
-    while (it_user_target != server->getClients().end())
-    {
-    	if (it_user_target->second.getNickname() == mode_infos.params)
-        	break;
-    	it_user_target++;
-    }
-	if (it_user_target == server->getClients().end())
-	{
-		addToClientBuffer(server, client_fd, ERR_NOSUCHNICK(it_client->second.getNickname(), mode_infos.params));
-		addToClientBuffer(server, client_fd, ERR_USERNOTINCHANNEL(it_client->second.getNickname(), mode_infos.params, mode_infos.target));
-		return ;
-	}
-	// User IRC mais checker qu'il est membre du channel
-	if (it_channel_target->second.getClientList().find(mode_infos.params) == it_channel_target->second.getClientList().end())
-	{
-		addToClientBuffer(server, client_fd, ERR_USERNOTINCHANNEL(it_client->second.getNickname(), mode_infos.params, mode_infos.target));
-		return ;
-	}
-
-	// Checker si param est dans la liste des operators du channel
-	std::vector<std::string>::iterator it;
-	for (it = it_channel_target->second.getOperators().begin(); it != it_channel_target->second.getOperators().end(); it++)
-	{
-		if (*it == mode_infos.params)
-			break;
-	}
-	// checker si c'est pour rajouter ou enlever le mode
-	if (str[0] == '+')
-	{
-		if (it != it_channel_target->second.getOperators().end()) // si param est deja operator
-			return ;
-		
-		// add au operator et send la reply en broadcast
-		it_channel_target->second.getOperators().push_back(mode_infos.params);
-		broadcastToAllChannelMembers(server, it_channel_target->second, MODE_CHANNELMSGWITHPARAM(mode_infos.target, "+o", mode_infos.params));
-	}
-	else // si c'est '-'
-	{
-		if (it == it_channel_target->second.getOperators().end()) // si param n'est pas operator
-			return ;
-		// retirer du vector et send la reply en broadcast
-		it_channel_target->second.getOperators().erase(it);
-		broadcastToAllChannelMembers(server, it_channel_target->second, MODE_CHANNELMSGWITHPARAM(mode_infos.target, "-o", mode_infos.params));
-	}
-	// std::cout << "check operators du channel :" << std::endl;
-	// for (std::vector<std::string>::iterator it = it_channel_target->second.getOperators().begin(); it != it_channel_target->second.getOperators().end(); it++)
-	// 	std::cout << *it << std::endl;
-}
-
-static void	topicChannelMode(Server *server, mode_struct mode_infos, int const client_fd, std::string mode_str)
-{
-	std::map<const int, Client>::iterator it_client = server->getClients().find(client_fd);
-	(void)it_client;
-	std::map<std::string, Channel>::iterator it_channel_target = server->getChannels().find(mode_infos.target);
-
-	size_t pos = it_channel_target->second.getMode().find("t");
-	if (mode_str[0] == '+')
-	{
-		if (pos != std::string::npos) // le mode est deja présent
-			return;
-		it_channel_target->second.addMode("t");
-		broadcastToAllChannelMembers(server, it_channel_target->second, MODE_CHANNELMSG(mode_infos.target, "+t"));
-	}
-	else // si c'est '-'
-	{
-		if (pos == std::string::npos) // le mode n'est pas présent
-			return;
-		it_channel_target->second.removeMode("t");
-		broadcastToAllChannelMembers(server, it_channel_target->second, MODE_CHANNELMSG(mode_infos.target, "-t"));
-	}
-}
-
-static void	secretChannelMode(Server *server, mode_struct mode_infos, int const client_fd, std::string mode_str)
-{
-	std::map<const int, Client>::iterator it_client = server->getClients().find(client_fd);
-	(void)it_client;
-	std::map<std::string, Channel>::iterator it_channel_target = server->getChannels().find(mode_infos.target);
-
-	size_t pos = it_channel_target->second.getMode().find("s");
-	if (mode_str[0] == '+')
-	{
-		if (pos != std::string::npos) // le mode est deja présent
-			return;
-		it_channel_target->second.addMode("s");
-		broadcastToAllChannelMembers(server, it_channel_target->second, MODE_CHANNELMSG(mode_infos.target, "+s"));
-	}
-	else // si c'est '-'
-	{
-		if (pos == std::string::npos) // le mode n'est pas présent
-			return;
-		it_channel_target->second.removeMode("s");
-		broadcastToAllChannelMembers(server, it_channel_target->second, MODE_CHANNELMSG(mode_infos.target, "-s"));
-	}
-}
-
-static bool isAlpha(std::string str) 
-{
-    for (size_t i = 0; i < str.length(); i++) 
-	{
-        if (!isalpha(str[i])) 
-            return (false);
-    }
-    return (true);
-}
-
-static void	keyChannelMode(Server *server, mode_struct mode_infos, int const client_fd, std::string mode_str)
-{
-	std::map<const int, Client>::iterator it_client = server->getClients().find(client_fd);
-	std::map<std::string, Channel>::iterator it_channel_target = server->getChannels().find(mode_infos.target);
-
-
-	size_t pos = it_channel_target->second.getMode().find("k");
-	if (mode_infos.params.empty() == true)
-		return ;
-	
-	if (mode_str[0] == '+')
-	{
-		if (pos != std::string::npos) // le mode est deja présent
-			return;
-		if (isAlpha(mode_infos.params) == false) // check si le mdp est conforme
-		{
-			addToClientBuffer(server, client_fd, ERR_INVALIDMODEPARAM(it_client->second.getNickname(), mode_infos.target, "+k", mode_infos.params));
-			return ;
-		}
-		it_channel_target->second.addMode("k");
-		it_channel_target->second.setChannelPassword(mode_infos.params);
-		broadcastToAllChannelMembers(server, it_channel_target->second, MODE_CHANNELMSGWITHPARAM(mode_infos.target, "+k", mode_infos.params));
-	}
-	else // si c'est '-'
-	{
-		if (pos == std::string::npos) // le mode n'est pas présent
-			return;
-		it_channel_target->second.removeMode("k");
-		it_channel_target->second.removeChannelPassword();
-		broadcastToAllChannelMembers(server, it_channel_target->second, MODE_CHANNELMSGWITHPARAM(mode_infos.target, "-k", mode_infos.params));
-	}
 }
 
 static void	changeChannelMode(Server *server, mode_struct mode_infos, int const client_fd)
